# lib/api-zod/src/generated/types/createWaitlistLeadRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateWaitlistLeadRequest {
  /** @minLength 1 */
  firstName: string;
  email: string;
  /** @minLength 1 */
  canton: string;
  /** @minLength 1 */
  situation: string;
}