# lib/api-zod/src/generated/types/waitlistLead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WaitlistLead {
  id: number;
  firstName: string;
  email: string;
  canton: string;
  situation: string;
  createdAt: Date;
}